// src/app/model/question3.ts
export interface Question3 {
    id: number;
    situation: string;
    video: string;
    quiz: {};
    // [
    //     {
    //         id: number,
    //         problem: string,
    //         option1: string,
    //         option2: string,
    //         option3: string,
    //         option4: string,
    //         answer: string
    //     }
    // ];
}